+++ desktop/gervill/src/main/java/media/sound/SoftEnvelopeGenerator.java
@@ -127,7 +127,7 @@ public class SoftEnvelopeGenerator implements SoftProcess {
             if ((stage[i] > EG_OFF) && (stage[i] < EG_RELEASE)) {
                 if (on[i][0] < 0.5) {
                     if (on[i][0] < -0.5) {
-                        stage_count[i] = (int)(Math.pow(2,
+                        stage_count[i] = (int) (Math.pow(2,
                                 this.shutdown[i][0] / 1200.0) / control_time);
                         if (stage_count[i] < 0)
                             stage_count[i] = 0;
@@ -143,17 +143,17 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                             continue;
                         }
 
-                        stage_count[i] = (int)(Math.pow(2,
+                        stage_count[i] = (int) (Math.pow(2,
                                 this.release[i][0] / 1200.0) / control_time);
                         stage_count[i]
-                                += (int)(this.release2[i][0]/(control_time * 1000));
+                                += (int) (this.release2[i][0]/(control_time * 1000));
                         if (stage_count[i] < 0)
                             stage_count[i] = 0;
                         // stage_v[i] = out[i][0];
                         stage_ix[i] = 0;
 
                         double m = 1 - out[i][0];
-                        stage_ix[i] = (int)(stage_count[i] * m);
+                        stage_ix[i] = (int) (stage_count[i] * m);
 
                         stage[i] = EG_RELEASE;
                     }
@@ -166,7 +166,7 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                 if (on[i][0] < 0.5)
                     break;
                 stage[i] = EG_DELAY;
-                stage_ix[i] = (int)(Math.pow(2,
+                stage_ix[i] = (int) (Math.pow(2,
                         this.delay[i][0] / 1200.0) / control_time);
                 if (stage_ix[i] < 0)
                     stage_ix[i] = 0;
@@ -179,14 +179,14 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                             && (attack < 0 && Double.isInfinite(attack))) {
                         out[i][0] = 1;
                         stage[i] = EG_HOLD;
-                        stage_count[i] = (int)(Math.pow(2,
+                        stage_count[i] = (int) (Math.pow(2,
                                 this.hold[i][0] / 1200.0) / control_time);
                         stage_ix[i] = 0;
                     } else {
                         stage[i] = EG_ATTACK;
-                        stage_count[i] = (int)(Math.pow(2,
+                        stage_count[i] = (int) (Math.pow(2,
                                 attack / 1200.0) / control_time);
-                        stage_count[i] += (int)(attack2 / (control_time * 1000));
+                        stage_count[i] += (int) (attack2 / (control_time * 1000));
                         if (stage_count[i] < 0)
                             stage_count[i] = 0;
                         stage_ix[i] = 0;
@@ -201,7 +201,7 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                     stage[i] = EG_HOLD;
                 } else {
                     // CONVEX attack
-                    double a = ((double)stage_ix[i]) / ((double)stage_count[i]);
+                    double a = (double) stage_ix[i] / stage_count[i];
                     a = 1 + ((40.0 / 96.0) / Math.log(10)) * Math.log(a);
                     if (a < 0)
                         a = 0;
@@ -214,9 +214,9 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                 stage_ix[i]++;
                 if (stage_ix[i] >= stage_count[i]) {
                     stage[i] = EG_DECAY;
-                    stage_count[i] = (int)(Math.pow(2,
+                    stage_count[i] = (int) (Math.pow(2,
                             this.decay[i][0] / 1200.0) / control_time);
-                    stage_count[i] += (int)(this.decay2[i][0]/(control_time*1000));
+                    stage_count[i] += (int) (this.decay2[i][0]/(control_time*1000));
                     if (stage_count[i] < 0)
                         stage_count[i] = 0;
                     stage_ix[i] = 0;
@@ -234,7 +234,7 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                         stage[i] = EG_END;
                     }
                 } else {
-                    double m = ((double)stage_ix[i]) / ((double)stage_count[i]);
+                    double m = (double) stage_ix[i] / stage_count[i];
                     out[i][0] = (1 - m) + sustain * m;
                 }
                 break;
@@ -247,11 +247,11 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                     active[i][0] = 0;
                     stage[i] = EG_END;
                 } else {
-                    double m = ((double)stage_ix[i]) / ((double)stage_count[i]);
+                    double m = (double) stage_ix[i] / stage_count[i];
                     out[i][0] = (1 - m); // *stage_v[i];
 
                     if (on[i][0] < -0.5) {
-                        stage_count[i] = (int)(Math.pow(2,
+                        stage_count[i] = (int) (Math.pow(2,
                                 this.shutdown[i][0] / 1200.0) / control_time);
                         if (stage_count[i] < 0)
                             stage_count[i] = 0;
@@ -265,10 +265,10 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                         sustain = this.sustain[i][0] * (1.0 / 1000.0);
                         if (out[i][0] > sustain) {
                             stage[i] = EG_DECAY;
-                            stage_count[i] = (int)(Math.pow(2,
+                            stage_count[i] = (int) (Math.pow(2,
                                     this.decay[i][0] / 1200.0) / control_time);
                             stage_count[i] +=
-                                    (int)(this.decay2[i][0]/(control_time*1000));
+                                    (int) (this.decay2[i][0]/(control_time*1000));
                             if (stage_count[i] < 0)
                                 stage_count[i] = 0;
                             m = (out[i][0] - 1) / (sustain - 1);
@@ -285,7 +285,7 @@ public class SoftEnvelopeGenerator implements SoftProcess {
                     active[i][0] = 0;
                     stage[i] = EG_END;
                 } else {
-                    double m = ((double)stage_ix[i]) / ((double)stage_count[i]);
+                    double m = (double) stage_ix[i] / stage_count[i];
                     out[i][0] = (1 - m) * stage_v[i];
                 }
                 break;
